+++ src/__mocks__/methods.js
@@ -1,0 +1,8 @@
+module.exports = [
+    'get',
+    'post',
+    'push',
+    'delete',
+    'options',
+    'head',
+].sort();