+++ src/server/nekoproxy.js
@@ -1,4 +1,6 @@
 import config from 'config';
-import { Server } from './server';
+import {Server} from './server';
+import {endpoint} from './metrics';
 
 export const server = new Server(config);
+endpoint(config.get('metrics'));