+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+# [Unreleased]
+### Added
+- Prometheus client.
+
 ## [1.0.0] - 2016-03-16
 ### Changed
 - Start following [Keep a CHANGELOG](http://keepachangelog.com/).